添加ufw防火墙判断

+++ bin/init.py
@@ -11,13 +11,20 @@ def main():
     log=Logger({"remote": log_remote_level}, logger_name="init")
     return_value=0
 
-    log.logger.info(f"关闭防火墙")
-    firewalld_cmd=f"systemctl disable firewalld && systemctl stop firewalld"
-    log.logger.debug(f"{firewalld_cmd=}")
-    result, msg=exec_command(firewalld_cmd)
-    if not result:
-        log.logger.error(f"关闭防火墙失败: {msg}")
-        return_value=error_code
+    if os.path.exists("/etc/firewalld/firewalld.conf"):
+        firewalld_cmd=f"systemctl disable firewalld && systemctl stop firewalld"
+    elif os.path.exists("/etc/default/ufw"):
+        firewalld_cmd=f"ufw disable"
+    else:
+        firewalld_cmd=""
+
+    if firewalld_cmd!="":
+        log.logger.info(f"关闭防火墙")
+        log.logger.debug(f"{firewalld_cmd=}")
+        result, msg=exec_command(firewalld_cmd)
+        if not result:
+            log.logger.error(f"关闭防火墙失败: {msg}")
+            return_value=error_code
 
     # 关闭selinux
     selinux_conf_file="/etc/selinux/config"
@@ -43,7 +50,7 @@ def main():
     value=65536
     ulimit_conf_file="/etc/security/limits.conf"
 
-    result, msg=exec_command("ulimit -n")
+    result, msg=exec_command("bash -c 'ulimit -n'")
     if result:
         nofile_value=int(msg)
         if nofile_value < value:
@@ -54,7 +61,7 @@ def main():
         log.logger.error(f"获取nofile失败: {msg}")
         return_value=error_code
 
-    result, msg=exec_command("ulimit -u")
+    result, msg=exec_command("bash -c 'ulimit -u'")
     if result:
         nproc_value=int(msg)
         if nproc_value < value:
